backend/requirements.txt
fastapi[standard]
python-multipart
uvicorn
starlette
pillow
cupy